# docs/swagger.yaml
basePath: /api/v1
definitions:
  errorHandling.AppError:
    properties:
      code:
        type: integer
      key:
        type: string
      log:
        type: string
      message:
        type: string
    type: object
...
  models.AppSetting:
    properties:
      Android:
        additionalProperties: true
        type: object
      Config:
        additionalProperties: true
        type: object
      Ios:
        additionalProperties: true
        type: object
      Key:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Category:
    properties:
      Image:
        type: string
      Name:
        type: string
      Products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      ProductsStore:
        items:
          $ref: '#/definitions/models.ProductStore'
        type: array
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Image:
    properties:
      Url:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  models.Product:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      ConditionVoucher:
        type: string
      Content:
        type: string
      Cost:
        type: number
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      HasPromote:
        type: boolean
      HasPromoteVoucher:
        type: boolean
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      Percent:
        type: number
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      ValueVoucher:
        type: number
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      Voucher:
        $ref: '#/definitions/models.Voucher'
        type: object
      VoucherId:
        $ref: '#/definitions/models.UID'
        type: object
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.ProductStore:
    properties:
      ActiveElement:
        type: string
      Avatar:
        type: string
      Barcode:
        type: string
      BasePrice:
        type: number
      Category:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      Code:
        type: string
      Content:
        type: string
      Description:
        type: string
      Direction:
        type: string
      DoNotUse:
        type: string
      DrugInteractions:
        type: string
      GlobalManufacturerName:
        type: string
      Images:
        items:
          $ref: '#/definitions/models.Image'
        type: array
      IndicationsOfTheDrug:
        type: string
      Manufacturer:
        type: string
      Name:
        type: string
      Overdose:
        type: string
      PackagingSize:
        type: string
      RegistrationNo:
        type: string
      Status:
        type: string
      Storage:
        type: string
      Tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      Unit:
        type: string
      Variants:
        items:
          $ref: '#/definitions/models.Variant'
        type: array
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.Promotion:
    properties:
      AreaId:
        $ref: '#/definitions/models.UID'
        type: object
      Avatar:
        type: string
      Deleted:
        type: boolean
      EndTime:
        type: integer
      Name:
        type: string
      Note:
        type: string
      StartTime:
        type: integer
      Status:
        type: boolean
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      promotionDetails:
        items:
          $ref: '#/definitions/models.PromotionDetail'
        type: array
      updated_at:
        type: integer
    type: object
  models.PromotionDetail:
    properties:
      Condition:
        type: string
      Percent:
        type: number
      Product:
        $ref: '#/definitions/models.Product'
        type: object
      ProductID:
        $ref: '#/definitions/models.UID'
        type: object
      Promotion:
        $ref: '#/definitions/models.Promotion'
        type: object
      PromotionID:
        $ref: '#/definitions/models.UID'
        type: object
      Type:
        type: string
      Value:
        type: number
      Variant:
        $ref: '#/definitions/models.Variant'
        type: object
      VariantID:
        $ref: '#/definitions/models.UID'
        type: object
      VoucherID:
        $ref: '#/definitions/models.UID'
        type: object
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
      voucher:
        $ref: '#/definitions/models.Voucher'
        type: object
    type: object
...
  models.Tag:
    properties:
      Name:
        type: string
      Slug:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
  models.UID:
    type: object
...
  models.Variant:
    properties:
      Name:
        type: string
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
      variantValue:
        items:
          $ref: '#/definitions/models.VariantValue'
        type: array
    type: object
  models.VariantValue:
    properties:
      convertValue:
        type: number
      operator:
        type: string
      product:
        $ref: '#/definitions/models.Product'
        type: object
      productID:
        type: integer
      variant:
        $ref: '#/definitions/models.Variant'
        type: object
      variantID:
        type: integer
    type: object
  models.Voucher:
    properties:
      Deleted:
        type: boolean
      MaxValue:
        type: number
      Name:
        type: string
      Note:
        type: string
      Type:
        type: string
      Unit:
        type: string
      Value:
        type: number
      created_at:
        type: integer
      id:
        $ref: '#/definitions/models.UID'
        type: object
      updated_at:
        type: integer
    type: object
...
  requests.SearchSettingRequest:
    properties:
      Key:
        type: string
    type: object
...
  requests.SearchVoucherRequest:
    properties:
      Name:
        type: string
      Type:
        type: string
      limit:
        example: 10
        type: integer
      offset:
        example: 0
        type: integer
      sort:
        $ref: '#/definitions/requests.SortOption'
        type: object
    type: object
...
  requests.SettingRequest:
    properties:
      Android:
        additionalProperties: true
        type: object
      Config:
        additionalProperties: true
        type: object
      Ios:
        additionalProperties: true
        type: object
      Key:
        type: string
    type: object
  requests.SortOption:
    properties:
      sort_direction:
        type: string
      sort_field:
        type: string
    type: object
...
  requests.VoucherRequest:
    properties:
      MaxValue:
        type: number
      Name:
        type: string
      Note:
        type: string
      Type:
        type: string
      Value:
        type: number
      unit:
        type: string
    type: object
...
  responses.Data:
    properties:
      code:
        type: integer
      key:
        type: string
      message:
        type: string
    type: object
...
  responses.VoucherSearch:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.Voucher'
        type: array
      message:
        type: string
      total:
        type: integer
    type: object
host: v1.api.medilane.vn
info:
  contact:
    name: medilane team
    url: https://www.medilane.vn/
  description: This is openapi for account api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Medilane account api
  version: "1.0"
paths:
...
  /setting:
    post:
      consumes:
      - application/json
      description: Perform create
      operationId: create-setting
      parameters:
      - description: Create setting
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AppSetting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create setting
      tags:
      - Setting Management
  /setting/{id}:
    put:
      consumes:
      - application/json
      description: Perform create
      operationId: edit-setting
      parameters:
      - description: Create setting
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SettingRequest'
      - description: id account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AppSetting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit setting
      tags:
      - Setting Management
  /setting/find:
    post:
      consumes:
      - application/json
      description: Perform search setting
      operationId: setting-promotion
      parameters:
      - description: Filter setting
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppSetting'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search setting in system
      tags:
      - Setting Management
...
  /voucher:
    post:
      consumes:
      - application/json
      description: Perform create voucher
      operationId: create-voucher
      parameters:
      - description: Create promotion
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.VoucherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Voucher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Create voucher  in system
      tags:
      - Voucher Management
  /voucher/{id}:
    delete:
      consumes:
      - application/json
      description: Perform delete voucher
      operationId: delete-voucher
      parameters:
      - description: id voucher
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Delete voucher (soft delete) in system
      tags:
      - Voucher Management
    get:
      consumes:
      - application/json
      description: Perform get voucher
      operationId: get-voucher
      parameters:
      - description: id voucher
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Voucher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Get voucher in system
      tags:
      - Voucher Management
    put:
      consumes:
      - application/json
      description: Perform edit voucher
      operationId: edit-voucher
      parameters:
      - description: body voucher
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.VoucherRequest'
      - description: id voucher
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Promotion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Edit voucher in system
      tags:
      - Voucher Management
  /voucher/find:
    post:
      consumes:
      - application/json
      description: Perform search voucher
      operationId: search-voucher
      parameters:
      - description: Filter voucher
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.SearchVoucherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.VoucherSearch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errorHandling.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorHandling.AppError'
      security:
      - BearerAuth: []
      summary: Search voucher in system
      tags:
      - Voucher Management
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"